just general refactor and removal of unused code
